Refactor namespaces for Categories and Reporting tests
Updated namespaces for `CategoriesDbFixture`, `IntegrationTests`, `ReportingWebAppFactory`, and `ReportingControllerTests` to improve organization and fix typos. This ensures all tests are correctly referencing their respective fixtures and maintains consistency across the codebase.
